Add mock data to Question component to simulate fetch request
