+++ src/Components/Question/Question.js
@@ -3,28 +3,28 @@ import "./Question.css"
 import Audio from '../Audio/Audio'
 import Choices from "../Choices/Choices";
 
-const Question = () => {
+const Question = ({deck}) => {
 
-  const deck = [
-    {
-      id: 16933,
-      audioURL: "https://cdn.freesound.org/previews/16/16933_37876-hq.mp3",
-      correctAnswer: "Donkey",
-      wrongAnswers: ["Horse", "Goat", "Cow"]
-    },
-    {
-      id: 73371,
-      audioURL: "https://cdn.freesound.org/previews/73/73371_238949-hq.mp3",
-      correctAnswer: "Bee",
-      wrongAnswers: ["Hornet", "Humming Bird", "Mosquito"]
-    },
-    {
-      id: 510917,
-      audioURL: "https://cdn.freesound.org/previews/510/510917_11157357-hq.mp3",
-      correctAnswer: "Seagull",
-      wrongAnswers: ["Pelican", "Eagle", "Hawk"],
-    }
-  ]  
+  // const deck = [
+  //   {
+  //     id: 16933,
+  //     audioURL: "https://cdn.freesound.org/previews/16/16933_37876-hq.mp3",
+  //     correctAnswer: "Donkey",
+  //     wrongAnswers: ["Horse", "Goat", "Cow"]
+  //   },
+  //   {
+  //     id: 73371,
+  //     audioURL: "https://cdn.freesound.org/previews/73/73371_238949-hq.mp3",
+  //     correctAnswer: "Bee",
+  //     wrongAnswers: ["Hornet", "Humming Bird", "Mosquito"]
+  //   },
+  //   {
+  //     id: 510917,
+  //     audioURL: "https://cdn.freesound.org/previews/510/510917_11157357-hq.mp3",
+  //     correctAnswer: "Seagull",
+  //     wrongAnswers: ["Pelican", "Eagle", "Hawk"],
+  //   }
+  // ]  
 
   const [turn, setTurn] = useState(0)
